services/ml-audit-service/requirements.txt
fastapi
uvicorn[standard]
requests
google-auth
python-dotenv
